Extract sortable item component.
